feat(weather): make DaylightUI to update its state according to current time of the day

+++ feature/weather/src/main/kotlin/prus/justweatherapp/feature/weather/location/current/daylight/DaylightUiModel.kt
@@ -2,6 +2,6 @@ package prus.justweatherapp.feature.weather.location.current.daylight
 
 data class DaylightUiModel(
     val text:String,
-    val percentage: Double,
+    val percentage: Float,
     val isDay:Boolean
 )
